requirements.txt
# requirements.txt
# This file is managed by pip-tools. Do not edit manually.  
torch>=2.1.0
deepspeed>=0.13.0
transformers>=4.36.0
fsspec[s3]>=2024.1.0
pynvml>=11.5.0
prometheus-client>=0.17.0
matplotlib>=3.7.2
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0
